# docs/swagger.yaml
definitions:
...
  usercenter.CommonResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  usercenter.DelRequest:
    properties:
      id:
        description: '@inject_tag: validate:"required"'
        type: string
      tenantID:
        type: string
    required:
    - id
    type: object
...
  usercenter.GetAllWechatConfigResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.WechatConfigInfo'
        type: array
      message:
        type: string
    type: object
...
  usercenter.GetSystemConfigDetailResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/usercenter.SystemConfigInfo'
      message:
        type: string
    type: object
...
  usercenter.GetWechatConfigDetailResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/usercenter.WechatConfigInfo'
      message:
        type: string
    type: object
...
  usercenter.QuerySystemConfigResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.SystemConfigInfo'
        type: array
      message:
        type: string
      pages:
        type: integer
      records:
        type: integer
      total:
        type: integer
    type: object
...
  usercenter.QueryWechatConfigResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.WechatConfigInfo'
        type: array
      message:
        type: string
      pages:
        type: integer
      records:
        type: integer
      total:
        type: integer
    type: object
...
  usercenter.SystemConfigInfo:
    properties:
      defaultPwd:
        description: 重置默认密码
        type: string
      enableTenant:
        description: 启用租户
        type: boolean
      expired:
        description: Token过期时间 单位分钟
        type: string
      id:
        type: string
      isMust:
        description: 系统必须要有的数据
        type: boolean
      key:
        description: 加密key
        type: string
      platformTenantID:
        description: 平台租户
        type: string
      projectID:
        type: string
      redisAddr:
        description: Redis地址
        type: string
      redisName:
        description: Redis用户名
        type: string
      redisPwd:
        description: Redis密码
        type: string
      superAdminRoleID:
        description: 超级管理员角色
        type: string
      tenantID:
        type: string
    type: object
...
  usercenter.WechatConfigInfo:
    properties:
      appID:
        description: APP ID
        type: string
      appName:
        description: APP名称
        type: string
      appType:
        description: 类型 1-微信小程序 2-微信公众号 3-微信APP应用 4-微信网站应用
        type: integer
      defaultRoleID:
        description: 默认角色 用户通过微信注册时赋予默认角色
        type: string
      displayName:
        description: 显示名称
        type: string
      encodingAESKey:
        description: 消息加解密密钥
        type: string
      encodingMethod:
        description: 消息加解密方式,1-明文模式,2-兼容模式,3-安全模式（推荐）
        type: integer
      id:
        type: string
      isMust:
        description: 系统必须要有的数据
        type: boolean
      projectID:
        type: string
      redirectUrl:
        description: 重定向URL
        type: string
      secret:
        description: 秘钥
        type: string
      tenantID:
        description: 租户ID
        type: string
      token:
        description: Token
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/core/system/config/add:
    post:
      consumes:
      - application/json
      description: 新增
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Add SystemConfig
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.SystemConfigInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 新增
      tags:
      - 系统配置管理
...
  /api/core/system/config/delete:
    delete:
      consumes:
      - application/json
      description: 删除
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Delete SystemConfig
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usercenter.DelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 删除
      tags:
      - 系统配置管理
  /api/core/system/config/detail:
    get:
      consumes:
      - application/json
      description: 查询明细
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetSystemConfigDetailResponse'
      summary: 查询明细
      tags:
      - 系统配置管理
  /api/core/system/config/query:
    get:
      consumes:
      - application/json
      description: 分页查询
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QuerySystemConfigResponse'
      summary: 分页查询
      tags:
      - 系统配置管理
  /api/core/system/config/update:
    put:
      consumes:
      - application/json
      description: 更新
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Update SystemConfig
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.SystemConfigInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 更新
      tags:
      - 系统配置管理
...
  /api/core/wechat/config/add:
    post:
      consumes:
      - application/json
      description: 新增
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Add WechatConfig
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.WechatConfigInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 新增
      tags:
      - 微信应用配置管理
  /api/core/wechat/config/all:
    get:
      consumes:
      - application/json
      description: 查询所有
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetAllWechatConfigResponse'
      summary: 查询所有
      tags:
      - 微信应用配置管理
  /api/core/wechat/config/delete:
    delete:
      consumes:
      - application/json
      description: 删除
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Delete WechatConfig
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usercenter.DelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 删除
      tags:
      - 微信应用配置管理
  /api/core/wechat/config/detail:
    get:
      consumes:
      - application/json
      description: 查询明细
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetWechatConfigDetailResponse'
      summary: 查询明细
      tags:
      - 微信应用配置管理
  /api/core/wechat/config/query:
    get:
      consumes:
      - application/json
      description: 分页查询
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: APP名称
        in: query
        name: appName
        type: string
      - description: 租户ID
        in: query
        name: tenantID
        type: string
      - description: 类型
        in: query
        name: appType
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryWechatConfigResponse'
      summary: 分页查询
      tags:
      - 微信应用配置管理
  /api/core/wechat/config/update:
    put:
      consumes:
      - application/json
      description: 更新
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Update WechatConfig
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/usercenter.WechatConfigInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 更新
      tags:
      - 微信应用配置管理
...
swagger: "2.0"